refactor(core): make markstream-core the shared diff preview implementation
Consolidate the diff-preview model (including the first/last-aware collapsed
'unmodified lines' pill) into markstream-core's buildDiffPreviewPanes and have
the Vue 3 PreCodeNode consume it, removing its inline copy.

markstream-core/src/diff-preview.ts is now feature-complete and the single
source of truth:
- collapsed rows carry collapsedFirst/collapsedLast flags (pierre separator
  parity: terminal rows reuse the pre's own bottom padding)
- 'No newline at end of file' metadata rows for patch and source diffs
- git file headers (diff --git / index / --- / +++ ) are dropped instead of
  painted as numbered context rows
- side-by-side patch rows are grouped and row-aligned with spacers, numbered
  from hunk headers
- full LCS matching (stable multi-hunk classification while streaming)

Vue 3 PreCodeNode now imports buildDiffPreviewPanes; its diff preview helpers
are deleted. Behavior for tested cases is unchanged (48 pre-code tests pass).
Core tests extended for the collapse flags and no-newline metadata.
