+++ packages/markstream-core/src/diff-preview.ts
@@ -1,11 +1,16 @@
-export type DiffPreviewLineKind = 'context' | 'removed' | 'added' | 'hunk' | 'collapsed' | 'spacer'
+export type DiffPreviewLineKind = 'context' | 'removed' | 'added' | 'hunk' | 'collapsed' | 'metadata' | 'spacer'
 
 export interface DiffPreviewLine {
   code: string
   empty: boolean
   key: string
   kind: DiffPreviewLineKind
+  metadataKind?: 'context' | 'removed' | 'added'
   number: number | string
+  /** True when this collapsed row replaces the very first diff rows. */
+  collapsedFirst?: boolean
+  /** True when this collapsed row extends to the end of the diff (terminal). */
+  collapsedLast?: boolean
 }
 
 export interface DiffPreviewPane {
@@ -37,6 +42,7 @@ interface SourceLineMatch {
 }
 
 const DIFF_HEADER_PREFIXES = ['diff ', 'index ', '--- ', '+++ ', '@@ ']
+const NO_NEWLINE_METADATA = '\\ No newline at end of file'
 
 function displaySource(source: unknown, loading: boolean) {
   const value = String(source ?? '')
@@ -57,7 +63,7 @@ function normalizeLanguage(language: unknown) {
 }
 
 function isBlank(code: string) {
-  return code.trim().length === 0
+  return String(code ?? '').trim().length === 0
 }
 
 function makeLine(
@@ -72,7 +78,7 @@ function makeLine(
     code,
     empty,
     key,
-    kind: empty && kind !== 'hunk' && kind !== 'collapsed' && kind !== 'spacer' && !preserveBlankKind ? 'context' : kind,
+    kind: empty && kind !== 'hunk' && kind !== 'spacer' && !preserveBlankKind ? 'context' : kind,
     number,
   }
 }
@@ -97,6 +103,52 @@ function normalizeDiffBody(body: string, headers: boolean) {
   return !headers && body.startsWith(' ') && !body.startsWith('  ') ? ` ${body}` : body
 }
 
+function hasFinalNewline(source: unknown) {
+  return /(?:\r\n|\n|\r)$/.test(String(source ?? ''))
+}
+
+function createMetadataLine(
+  key: string,
+  metadataKind: 'context' | 'removed' | 'added',
+): DiffPreviewLine {
+  return {
+    ...makeLine('No newline at end of file', 'metadata', key, ''),
+    metadataKind,
+  }
+}
+
+function appendInlineSourceMetadata(
+  lines: DiffPreviewLine[],
+  originalSource: unknown,
+  modifiedSource: unknown,
+) {
+  const originalMissing = String(originalSource ?? '').length > 0 && !hasFinalNewline(originalSource)
+  const modifiedMissing = String(modifiedSource ?? '').length > 0 && !hasFinalNewline(modifiedSource)
+  if (!originalMissing && !modifiedMissing)
+    return lines
+
+  if (originalMissing)
+    lines.push(createMetadataLine('inline-no-newline-original', 'removed'))
+  if (modifiedMissing)
+    lines.push(createMetadataLine('inline-no-newline-modified', 'added'))
+  return lines
+}
+
+function isExplicitDiffLanguage(language: unknown, raw: unknown) {
+  if (normalizeLanguage(language) === 'diff')
+    return true
+
+  const firstLine = String(raw ?? '').split(/\r?\n/, 1)[0]?.trim() ?? ''
+  return /^`{3,}\s*diff(?:\s|$)|^~{3,}\s*diff(?:\s|$)/.test(firstLine)
+}
+
+function hasPatchLines(lines: string[], language: unknown, raw: unknown) {
+  const hasRemoved = lines.some(line => isRemovedLine(line))
+  const hasAdded = lines.some(line => isAddedLine(line))
+  return (hasRemoved && hasAdded)
+    || (isExplicitDiffLanguage(language, raw) && (hasRemoved || hasAdded))
+}
+
 function computeMatches(original: string[], modified: string[]): SourceLineMatch[] {
   const prefix: SourceLineMatch[] = []
   let start = 0
@@ -116,7 +168,11 @@ function computeMatches(original: string[], modified: string[]): SourceLineMatch
 
   const originalLength = originalEnd - start + 1
   const modifiedLength = modifiedEnd - start + 1
-  if (originalLength <= 0 || modifiedLength <= 0 || (originalLength + 1) * (modifiedLength + 1) > 1_500_000)
+  if (originalLength <= 0 || modifiedLength <= 0)
+    return prefix.concat(suffix)
+
+  const maxCells = 1_500_000
+  if ((originalLength + 1) * (modifiedLength + 1) > maxCells)
     return prefix.concat(suffix)
 
   const columns = modifiedLength + 1
@@ -155,25 +211,181 @@ function computeMatches(original: string[], modified: string[]): SourceLineMatch
   return prefix.concat(middle, suffix)
 }
 
-function buildSourcePanes(original: string[], modified: string[]) {
+const PATCH_HEADER_PREFIXES = ['diff ', 'index ', '--- ', '+++ ']
+
+function isPatchHeaderLine(line: string, headers: boolean) {
+  return headers && PATCH_HEADER_PREFIXES.some(prefix => line.startsWith(prefix))
+}
+
+function buildInlinePatchPreviewLines(lines: string[]): DiffPreviewLine[] {
+  const result: DiffPreviewLine[] = []
+  let originalLine = 1
+  let modifiedLine = 1
+  const headers = hasDiffHeaders(lines)
+
+  for (const [index, raw] of lines.entries()) {
+    if (raw === NO_NEWLINE_METADATA) {
+      const previousKind = result.at(-1)?.kind
+      const metadataKind = previousKind === 'removed' || previousKind === 'added'
+        ? previousKind
+        : 'context'
+      result.push(createMetadataLine(`inline-no-newline-${index}`, metadataKind))
+    }
+    else if (isPatchHeaderLine(raw, headers)) {
+      continue
+    }
+    else if (raw.startsWith('@@')) {
+      const match = raw.match(/^@@\s+-(\d+)(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/)
+      if (match) {
+        originalLine = Number(match[1])
+        modifiedLine = Number(match[2])
+      }
+      result.push({
+        ...makeLine(raw, 'hunk', `inline-hunk-${index}`, ''),
+      })
+    }
+    else if (isRemovedLine(raw)) {
+      result.push({
+        ...makeLine(normalizeDiffBody(raw.slice(1), headers), 'removed', `inline-removed-${index}`, originalLine++, true),
+      })
+    }
+    else if (isAddedLine(raw)) {
+      result.push({
+        ...makeLine(normalizeDiffBody(raw.slice(1), headers), 'added', `inline-added-${index}`, modifiedLine++, true),
+      })
+    }
+    else {
+      const code = headers && raw.startsWith(' ') ? raw.slice(1) : raw
+      result.push({
+        ...makeLine(code, 'context', `inline-context-${index}`, modifiedLine),
+      })
+      originalLine++
+      modifiedLine++
+    }
+  }
+
+  return result
+}
+
+function buildInlineSourcePreviewLines(
+  originalSource: unknown,
+  modifiedSource: unknown,
+  loading: boolean,
+): DiffPreviewLine[] {
+  const original = splitSource(originalSource, loading)
+  const modified = splitSource(modifiedSource, loading)
+  const matches = computeMatches(original, modified)
+  if (matches.length > 0) {
+    const result: DiffPreviewLine[] = []
+    let originalIndex = 0
+    let modifiedIndex = 0
+
+    for (const match of matches) {
+      while (originalIndex < match.originalIndex) {
+        result.push({
+          ...makeLine(original[originalIndex], 'removed', `inline-removed-source-${originalIndex}`, originalIndex + 1, shouldPreserveBlankKind(original, originalIndex)),
+        })
+        originalIndex++
+      }
+      while (modifiedIndex < match.modifiedIndex) {
+        result.push({
+          ...makeLine(modified[modifiedIndex], 'added', `inline-added-source-${modifiedIndex}`, modifiedIndex + 1, shouldPreserveBlankKind(modified, modifiedIndex)),
+        })
+        modifiedIndex++
+      }
+      result.push({
+        ...makeLine(modified[match.modifiedIndex], 'context', `inline-context-source-${match.originalIndex}-${match.modifiedIndex}`, match.modifiedIndex + 1),
+      })
+      originalIndex = match.originalIndex + 1
+      modifiedIndex = match.modifiedIndex + 1
+    }
+
+    while (originalIndex < original.length) {
+      result.push({
+        ...makeLine(original[originalIndex], 'removed', `inline-removed-source-${originalIndex}`, originalIndex + 1, shouldPreserveBlankKind(original, originalIndex)),
+      })
+      originalIndex++
+    }
+    while (modifiedIndex < modified.length) {
+      result.push({
+        ...makeLine(modified[modifiedIndex], 'added', `inline-added-source-${modifiedIndex}`, modifiedIndex + 1, shouldPreserveBlankKind(modified, modifiedIndex)),
+      })
+      modifiedIndex++
+    }
+
+    return result
+  }
+
+  const result: DiffPreviewLine[] = []
+  let start = 0
+  let originalEnd = original.length - 1
+  let modifiedEnd = modified.length - 1
+
+  while (start <= originalEnd && start <= modifiedEnd && original[start] === modified[start]) {
+    result.push({
+      ...makeLine(modified[start], 'context', `inline-prefix-${start}`, start + 1),
+    })
+    start++
+  }
+
+  const suffix: DiffPreviewLine[] = []
+  while (originalEnd >= start && modifiedEnd >= start && original[originalEnd] === modified[modifiedEnd]) {
+    suffix.unshift({
+      ...makeLine(modified[modifiedEnd], 'context', `inline-suffix-${modifiedEnd}`, modifiedEnd + 1),
+    })
+    originalEnd--
+    modifiedEnd--
+  }
+
+  for (let index = start; index <= originalEnd; index++) {
+    result.push({
+      ...makeLine(original[index], 'removed', `inline-removed-source-${index}`, index + 1, shouldPreserveBlankKind(original, index)),
+    })
+  }
+
+  for (let index = start; index <= modifiedEnd; index++) {
+    result.push({
+      ...makeLine(modified[index], 'added', `inline-added-source-${index}`, index + 1, shouldPreserveBlankKind(modified, index)),
+    })
+  }
+
+  return result.concat(suffix)
+}
+
+function buildSideBySideSourcePreviewPanes(
+  originalSource: unknown,
+  modifiedSource: unknown,
+  loading: boolean,
+  hideUnchangedRegions: BuildDiffPreviewOptions['hideUnchangedRegions'],
+): DiffPreviewPane[] {
+  const originalSourceLines = splitSource(originalSource, loading)
+  const modifiedSourceLines = splitSource(modifiedSource, loading)
+  const matches = computeMatches(originalSourceLines, modifiedSourceLines)
   const originalLines: DiffPreviewLine[] = []
   const modifiedLines: DiffPreviewLine[] = []
-  const matches = computeMatches(original, modified)
   let originalIndex = 0
   let modifiedIndex = 0
   let blockIndex = 0
 
-  const appendChangedRows = (originalEnd: number, modifiedEnd: number) => {
-    const count = Math.max(originalEnd - originalIndex, modifiedEnd - modifiedIndex)
-    for (let offset = 0; offset < count; offset++) {
-      const nextOriginal = originalIndex + offset
-      const nextModified = modifiedIndex + offset
-      originalLines.push(nextOriginal < originalEnd
-        ? makeLine(original[nextOriginal], 'removed', `original-changed-${blockIndex}-${nextOriginal}`, nextOriginal + 1, shouldPreserveBlankKind(original, nextOriginal))
-        : makeLine('', 'spacer', `original-spacer-${blockIndex}-${offset}`, ''))
-      modifiedLines.push(nextModified < modifiedEnd
-        ? makeLine(modified[nextModified], 'added', `modified-changed-${blockIndex}-${nextModified}`, nextModified + 1, shouldPreserveBlankKind(modified, nextModified))
-        : makeLine('', 'spacer', `modified-spacer-${blockIndex}-${offset}`, ''))
+  const appendChangedBlock = (originalEnd: number, modifiedEnd: number) => {
+    const rowCount = Math.max(originalEnd - originalIndex, modifiedEnd - modifiedIndex)
+    for (let offset = 0; offset < rowCount; offset++) {
+      const nextOriginalIndex = originalIndex + offset
+      const nextModifiedIndex = modifiedIndex + offset
+      originalLines.push(nextOriginalIndex < originalEnd
+        ? {
+            ...makeLine(originalSourceLines[nextOriginalIndex], 'removed', `original-changed-${blockIndex}-${nextOriginalIndex}`, nextOriginalIndex + 1, shouldPreserveBlankKind(originalSourceLines, nextOriginalIndex)),
+          }
+        : {
+            ...makeLine('', 'spacer', `original-spacer-${blockIndex}-${offset}`, ''),
+          })
+      modifiedLines.push(nextModifiedIndex < modifiedEnd
+        ? {
+            ...makeLine(modifiedSourceLines[nextModifiedIndex], 'added', `modified-changed-${blockIndex}-${nextModifiedIndex}`, nextModifiedIndex + 1, shouldPreserveBlankKind(modifiedSourceLines, nextModifiedIndex)),
+          }
+        : {
+            ...makeLine('', 'spacer', `modified-spacer-${blockIndex}-${offset}`, ''),
+          })
     }
     originalIndex = originalEnd
     modifiedIndex = modifiedEnd
@@ -181,170 +393,129 @@ function buildSourcePanes(original: string[], modified: string[]) {
   }
 
   for (const match of matches) {
-    appendChangedRows(match.originalIndex, match.modifiedIndex)
-    originalLines.push(makeLine(original[match.originalIndex], 'context', `original-context-${match.originalIndex}-${match.modifiedIndex}`, match.originalIndex + 1))
-    modifiedLines.push(makeLine(modified[match.modifiedIndex], 'context', `modified-context-${match.originalIndex}-${match.modifiedIndex}`, match.modifiedIndex + 1))
+    appendChangedBlock(match.originalIndex, match.modifiedIndex)
+    originalLines.push({
+      ...makeLine(originalSourceLines[match.originalIndex], 'context', `original-context-${match.originalIndex}-${match.modifiedIndex}`, match.originalIndex + 1),
+    })
+    modifiedLines.push({
+      ...makeLine(modifiedSourceLines[match.modifiedIndex], 'context', `modified-context-${match.originalIndex}-${match.modifiedIndex}`, match.modifiedIndex + 1),
+    })
     originalIndex = match.originalIndex + 1
     modifiedIndex = match.modifiedIndex + 1
   }
-  appendChangedRows(original.length, modified.length)
+  appendChangedBlock(originalSourceLines.length, modifiedSourceLines.length)
 
-  return [
-    { key: 'original', className: 'markstream-pre__diff-pane--original', lines: originalLines },
-    { key: 'modified', className: 'markstream-pre__diff-pane--modified', lines: modifiedLines },
-  ] satisfies DiffPreviewPane[]
-}
-
-function buildInlineSourceLines(original: string[], modified: string[]) {
-  const panes = buildSourcePanes(original, modified)
-  const lines: DiffPreviewLine[] = []
-  for (let index = 0; index < panes[0].lines.length; index++) {
-    const originalLine = panes[0].lines[index]
-    const modifiedLine = panes[1].lines[index]
-    if (originalLine.kind === 'removed')
-      lines.push({ ...originalLine, key: `inline-${originalLine.key}` })
-    if (modifiedLine.kind !== 'spacer')
-      lines.push({ ...modifiedLine, key: `inline-${modifiedLine.key}` })
-  }
-  return lines
-}
-
-function readHunkStart(line: string) {
-  const match = line.match(/^@@\s+-(\d+)(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/)
-  return match ? { original: Number(match[1]), modified: Number(match[2]) } : undefined
-}
-
-function buildPatchPanes(lines: string[], inline: boolean) {
-  const original: DiffPreviewLine[] = []
-  const modified: DiffPreviewLine[] = []
-  const inlineLines: DiffPreviewLine[] = []
-  const headers = hasDiffHeaders(lines)
-  let originalNumber = 1
-  let modifiedNumber = 1
-  let index = 0
-
-  while (index < lines.length) {
-    const raw = lines[index]
-    if (headers && ['diff ', 'index ', '--- ', '+++ '].some(prefix => raw.startsWith(prefix))) {
-      index++
-      continue
-    }
-    if (raw.startsWith('@@')) {
-      const start = readHunkStart(raw)
-      if (start) {
-        originalNumber = start.original
-        modifiedNumber = start.modified
-      }
-      const originalHunk = makeLine(raw, 'hunk', `original-hunk-${index}`, '')
-      const modifiedHunk = makeLine(raw, 'hunk', `modified-hunk-${index}`, '')
-      original.push(originalHunk)
-      modified.push(modifiedHunk)
-      inlineLines.push({ ...modifiedHunk, key: `inline-hunk-${index}` })
-      index++
-      continue
-    }
-
-    if (isRemovedLine(raw) || isAddedLine(raw)) {
-      const removed: string[] = []
-      const added: string[] = []
-      const blockStart = index
-      while (index < lines.length && (isRemovedLine(lines[index]) || isAddedLine(lines[index]))) {
-        if (isRemovedLine(lines[index]))
-          removed.push(normalizeDiffBody(lines[index].slice(1), headers))
-        else
-          added.push(normalizeDiffBody(lines[index].slice(1), headers))
-        index++
-      }
-      const count = Math.max(removed.length, added.length)
-      for (let offset = 0; offset < count; offset++) {
-        const removedLine = offset < removed.length
-          ? makeLine(removed[offset], 'removed', `original-removed-${blockStart}-${offset}`, originalNumber++, true)
-          : makeLine('', 'spacer', `original-spacer-${blockStart}-${offset}`, '')
-        const addedLine = offset < added.length
-          ? makeLine(added[offset], 'added', `modified-added-${blockStart}-${offset}`, modifiedNumber++, true)
-          : makeLine('', 'spacer', `modified-spacer-${blockStart}-${offset}`, '')
-        original.push(removedLine)
-        modified.push(addedLine)
-        if (removedLine.kind !== 'spacer')
-          inlineLines.push({ ...removedLine, key: `inline-${removedLine.key}` })
-        if (addedLine.kind !== 'spacer')
-          inlineLines.push({ ...addedLine, key: `inline-${addedLine.key}` })
-      }
-      continue
-    }
-
-    const code = headers && raw.startsWith(' ') ? raw.slice(1) : raw
-    const originalLine = makeLine(code, 'context', `original-context-${index}`, originalNumber++)
-    const modifiedLine = makeLine(code, 'context', `modified-context-${index}`, modifiedNumber++)
-    original.push(originalLine)
-    modified.push(modifiedLine)
-    inlineLines.push({ ...modifiedLine, key: `inline-context-${index}` })
-    index++
+  const originalMissing = String(originalSource ?? '').length > 0 && !hasFinalNewline(originalSource)
+  const modifiedMissing = String(modifiedSource ?? '').length > 0 && !hasFinalNewline(modifiedSource)
+  if (originalMissing || modifiedMissing) {
+    originalLines.push(originalMissing
+      ? createMetadataLine('original-no-newline', 'removed')
+      : {
+          ...makeLine('', 'spacer', 'original-no-newline-spacer', ''),
+        })
+    modifiedLines.push(modifiedMissing
+      ? createMetadataLine('modified-no-newline', 'added')
+      : {
+          ...makeLine('', 'spacer', 'modified-no-newline-spacer', ''),
+        })
   }
 
-  return inline
-    ? [{ key: 'inline', className: 'markstream-pre__diff-pane--inline', lines: inlineLines }]
-    : [
-        { key: 'original', className: 'markstream-pre__diff-pane--original', lines: original },
-        { key: 'modified', className: 'markstream-pre__diff-pane--modified', lines: modified },
-      ]
+  return collapseDiffPanes([
+    {
+      key: 'original',
+      className: 'markstream-pre__diff-pane--original',
+      lines: originalLines,
+    },
+    {
+      key: 'modified',
+      className: 'markstream-pre__diff-pane--modified',
+      lines: modifiedLines,
+    },
+  ], hideUnchangedRegions)
 }
 
-function collapsePanes(
+function resolveCollapseOptions(value: BuildDiffPreviewOptions['hideUnchangedRegions']) {
+  if (value == null || value === false)
+    return null
+  const options = value === true ? {} : value
+  if (options.enabled === false)
+    return null
+  return {
+    contextLineCount: Math.max(0, Math.floor(options.contextLineCount ?? 2)),
+    minimumLineCount: Math.max(1, Math.floor(options.minimumLineCount ?? 4)),
+  }
+}
+
+function collapseDiffPanes(
   panes: DiffPreviewPane[],
-  value: BuildDiffPreviewOptions['hideUnchangedRegions'],
+  hideUnchangedRegions: BuildDiffPreviewOptions['hideUnchangedRegions'],
 ) {
-  if (!value || panes.length < 1 || panes.length > 2)
+  const options = resolveCollapseOptions(hideUnchangedRegions)
+  if (!options || panes.length < 1 || panes.length > 2)
     return panes
   if (panes.length === 2 && panes[0].lines.length !== panes[1].lines.length)
     return panes
 
-  const options = value === true ? {} : value
-  if (options.enabled === false)
-    return panes
-  const contextLineCount = Math.max(0, Math.floor(options.contextLineCount ?? 2))
-  const minimumLineCount = Math.max(1, Math.floor(options.minimumLineCount ?? 4))
   const original = panes[0].lines
   const modified = panes[1]?.lines
-  const isMatchingContext = (line: number) => !modified
-    || (modified[line].kind === 'context' && original[line].code === modified[line].code)
-  const ranges: Array<{ end: number, start: number }> = []
+  let sourceLineCount = original.length
+  while (
+    sourceLineCount > 0
+    && panes.every(pane => pane.lines[sourceLineCount - 1].kind === 'metadata')
+  ) {
+    sourceLineCount--
+  }
+  const isUnchangedRow = (lineIndex: number) => original[lineIndex].kind === 'context'
+    && (
+      modified === undefined
+      || (
+        modified[lineIndex].kind === 'context'
+        && original[lineIndex].code === modified[lineIndex].code
+      )
+    )
+  const collapsedRanges: Array<{ start: number, end: number, first: boolean, last: boolean }> = []
   let index = 0
-
-  while (index < original.length) {
+  while (index < sourceLineCount) {
     const start = index
-    while (
-      index < original.length
-      && original[index].kind === 'context'
-      && isMatchingContext(index)
-    ) {
+    while (index < sourceLineCount && isUnchangedRow(index)) {
       index++
     }
     const end = index
-    if (end - start >= minimumLineCount) {
-      const hiddenStart = start + (start === 0 ? 0 : contextLineCount)
-      const hiddenEnd = end - (end === original.length ? 0 : contextLineCount)
-      if (hiddenEnd - hiddenStart >= minimumLineCount)
-        ranges.push({ start: hiddenStart, end: hiddenEnd })
+    const runLength = end - start
+    if (runLength >= options.minimumLineCount) {
+      const hiddenStart = start + (start === 0 ? 0 : options.contextLineCount)
+      const isTerminalRange = end === sourceLineCount
+      const hiddenEnd = end - (isTerminalRange ? 0 : options.contextLineCount)
+      if (hiddenEnd - hiddenStart >= options.minimumLineCount) {
+        collapsedRanges.push({
+          start: hiddenStart,
+          end: isTerminalRange ? original.length : hiddenEnd,
+          first: hiddenStart === 0,
+          last: isTerminalRange,
+        })
+      }
     }
     if (index === start)
       index++
   }
 
-  if (!ranges.length)
+  if (!collapsedRanges.length)
     return panes
 
   return panes.map((pane, paneIndex) => {
     const lines: DiffPreviewLine[] = []
     let sourceIndex = 0
-    for (const range of ranges) {
+    for (const range of collapsedRanges) {
       lines.push(...pane.lines.slice(sourceIndex, range.start))
-      lines.push(makeLine(
-        paneIndex === 0 || panes.length === 1 ? `${range.end - range.start} unmodified lines` : '',
-        'collapsed',
-        `${pane.key}-collapsed-${range.start}-${range.end}`,
-        '',
-      ))
+      lines.push({
+        code: paneIndex === 0 ? 'Unmodified lines' : '',
+        kind: 'collapsed',
+        empty: false,
+        key: `${pane.key}-collapsed-${range.start}-${range.end}`,
+        number: '',
+        collapsedFirst: range.first,
+        collapsedLast: range.last,
+      })
       sourceIndex = range.end
     }
     lines.push(...pane.lines.slice(sourceIndex))
@@ -355,26 +526,147 @@ function collapsePanes(
 export function buildDiffPreviewPanes(options: BuildDiffPreviewOptions): DiffPreviewPane[] {
   const loading = options.loading === true
   const inline = options.inline === true
-  const original = splitSource(options.originalCode, loading)
-  const modified = splitSource(options.updatedCode, loading)
-  const hasSourcePair = options.originalCode != null || options.updatedCode != null
+  const codeLines = splitSource(options.code, loading)
+  const originalCode = options.originalCode
+  const updatedCode = options.updatedCode
+  const hasSourcePair = originalCode != null || updatedCode != null
 
-  let panes: DiffPreviewPane[]
-  if (hasSourcePair) {
-    panes = inline
-      ? [{ key: 'inline', className: 'markstream-pre__diff-pane--inline', lines: buildInlineSourceLines(original, modified) }]
-      : buildSourcePanes(original, modified)
-  }
-  else {
-    const patchLines = splitSource(options.code, loading)
-    const language = normalizeLanguage(options.language)
-    const firstLine = String(options.raw ?? '').split(/\r?\n/, 1)[0]?.trim() ?? ''
-    const explicitDiff = language === 'diff' || /^`{3,}\s*diff(?:\s|$)|^~{3,}\s*diff(?:\s|$)/.test(firstLine)
-    const hasPatch = patchLines.some(isRemovedLine) && patchLines.some(isAddedLine)
-    if (!hasPatch && !explicitDiff)
-      return []
-    panes = buildPatchPanes(patchLines, inline)
+  if (inline) {
+    const lines = hasSourcePair
+      ? appendInlineSourceMetadata(
+          buildInlineSourcePreviewLines(originalCode, updatedCode, loading),
+          originalCode,
+          updatedCode,
+        )
+      : buildInlinePatchPreviewLines(codeLines)
+
+    return collapseDiffPanes([
+      {
+        key: 'inline',
+        className: 'markstream-pre__diff-pane--inline',
+        lines,
+      },
+    ], options.hideUnchangedRegions)
   }
 
-  return collapsePanes(panes, options.hideUnchangedRegions)
+  if (!hasPatchLines(codeLines, options.language, options.raw) && hasSourcePair) {
+    return buildSideBySideSourcePreviewPanes(
+      originalCode,
+      updatedCode,
+      loading,
+      options.hideUnchangedRegions,
+    )
+  }
+
+  const { original, modified } = buildSplitPatchPanes(codeLines)
+
+  return collapseDiffPanes([
+    {
+      key: 'original',
+      className: 'markstream-pre__diff-pane--original',
+      lines: original.map((line, index) => ({
+        ...line,
+        key: `original-${index}`,
+        number: line.kind === 'metadata' || line.kind === 'spacer' ? '' : line.number,
+      })),
+    },
+    {
+      key: 'modified',
+      className: 'markstream-pre__diff-pane--modified',
+      lines: modified.map((line, index) => ({
+        ...line,
+        key: `modified-${index}`,
+        number: line.kind === 'metadata' || line.kind === 'spacer' ? '' : line.number,
+      })),
+    },
+  ], options.hideUnchangedRegions)
+}
+
+function readHunkStart(line: string) {
+  const match = line.match(/^@@\s+-(\d+)(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/)
+  return match ? { original: Number(match[1]), modified: Number(match[2]) } : undefined
+}
+
+/**
+ * Parse a patch into side-by-side panes. Removed/added runs are grouped and
+ * row-aligned with spacer rows, patch line numbers are tracked from hunk
+ * headers, and git file metadata headers (`diff --git`, `index`, `---`, `+++`)
+ * are dropped instead of painted as numbered context rows.
+ */
+function buildSplitPatchPanes(lines: string[]) {
+  const original: DiffPreviewLine[] = []
+  const modified: DiffPreviewLine[] = []
+  const headers = hasDiffHeaders(lines)
+  let originalNumber = 1
+  let modifiedNumber = 1
+  let previousKind: 'context' | 'removed' | 'added' = 'context'
+  let index = 0
+
+  while (index < lines.length) {
+    const raw = lines[index]
+    if (raw === NO_NEWLINE_METADATA) {
+      if (previousKind === 'removed') {
+        original.push(createMetadataLine('original-patch-no-newline', 'removed'))
+        modified.push(makeLine('', 'spacer', '', ''))
+      }
+      else if (previousKind === 'added') {
+        original.push(makeLine('', 'spacer', '', ''))
+        modified.push(createMetadataLine('modified-patch-no-newline', 'added'))
+      }
+      else {
+        original.push(createMetadataLine('original-patch-no-newline', 'context'))
+        modified.push(createMetadataLine('modified-patch-no-newline', 'context'))
+      }
+      index++
+      continue
+    }
+    if (isPatchHeaderLine(raw, headers)) {
+      index++
+      continue
+    }
+    if (raw.startsWith('@@')) {
+      const start = readHunkStart(raw)
+      if (start) {
+        originalNumber = start.original
+        modifiedNumber = start.modified
+      }
+      original.push(makeLine(raw, 'hunk', '', ''))
+      modified.push(makeLine(raw, 'hunk', '', ''))
+      previousKind = 'context'
+      index++
+      continue
+    }
+    if (isRemovedLine(raw) || isAddedLine(raw)) {
+      const removed: string[] = []
+      const added: string[] = []
+      while (index < lines.length && (isRemovedLine(lines[index]) || isAddedLine(lines[index]))) {
+        if (isRemovedLine(lines[index])) {
+          removed.push(normalizeDiffBody(lines[index].slice(1), headers))
+          previousKind = 'removed'
+        }
+        else {
+          added.push(normalizeDiffBody(lines[index].slice(1), headers))
+          previousKind = 'added'
+        }
+        index++
+      }
+      const count = Math.max(removed.length, added.length)
+      for (let offset = 0; offset < count; offset++) {
+        original.push(offset < removed.length
+          ? makeLine(removed[offset], 'removed', '', originalNumber++, true)
+          : makeLine('', 'spacer', '', ''))
+        modified.push(offset < added.length
+          ? makeLine(added[offset], 'added', '', modifiedNumber++, true)
+          : makeLine('', 'spacer', '', ''))
+      }
+      continue
+    }
+    const code = headers && raw.startsWith(' ') ? raw.slice(1) : raw
+    original.push(makeLine(code, 'context', '', originalNumber++))
+    modified.push(makeLine(code, 'context', '', modifiedNumber++))
+    previousKind = 'context'
+    index++
+  }
+
+  return { original, modified }
 }
