+++ src/components/PreCodeNode/PreCodeNode.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
-import type { CodeBlockDiffHideUnchangedRegionsOptions, PreCodeNodeProps } from '../../types/component-props'
+import type { PreCodeNodeProps } from '../../types/component-props'
 import type { DiffLineMetric } from './preCodeDiffMetrics'
+import { buildDiffPreviewPanes } from 'markstream-core'
 
 import { computed, normalizeClass, normalizeStyle, onBeforeUnmount, ref, useAttrs, watch } from 'vue'
 
@@ -54,9 +55,6 @@ function countCodeLines(code: string) {
 }
 
 const codeLineCount = computed(() => countCodeLines(displayCode.value))
-const codeLines = computed(() => {
-  return displayCode.value.split(/\r\n|\n|\r/)
-})
 const logicalCodeLines = computed(() => {
   const code = displayCode.value
   const lines = code.split(/(?<=\n)|(?<=\r)(?!\n)/)
@@ -122,54 +120,6 @@ const reservedHeightStyle = computed(() => {
   }
 })
 
-type DiffPreviewLineKind = 'context' | 'removed' | 'added' | 'hunk' | 'collapsed' | 'metadata' | 'spacer'
-const DIFF_HEADER_PREFIXES = ['diff ', 'index ', '--- ', '+++ ', '@@ ']
-const NO_NEWLINE_METADATA = '\\ No newline at end of file'
-
-interface DiffPreviewLine {
-  code: string
-  kind: DiffPreviewLineKind
-  metadataKind?: 'context' | 'removed' | 'added'
-  empty: boolean
-  key: string
-  number: number | string
-  /** True when this collapsed row replaces the very first diff rows. */
-  collapsedFirst?: boolean
-  /** True when this collapsed row extends to the end of the diff (terminal). */
-  collapsedLast?: boolean
-}
-
-interface DiffPreviewPane {
-  key: string
-  className: string
-  lines: DiffPreviewLine[]
-}
-
-interface SourceLineMatch {
-  originalIndex: number
-  modifiedIndex: number
-}
-
-function isBlankDiffPreviewLine(code: string) {
-  return String(code ?? '').trim().length === 0
-}
-
-function toDiffPreviewLine(
-  code: string,
-  kind: DiffPreviewLineKind = 'context',
-  options: { preserveBlankKind?: boolean } = {},
-) {
-  const empty = isBlankDiffPreviewLine(code)
-  return {
-    code,
-    // Do not paint terminal blank lines / visual spacer rows as added/removed.
-    // This keeps the pre fallback close to stream-diffs, where the empty continuation
-    // surface should not flash red/green before highlighting is ready.
-    kind: empty && kind !== 'hunk' && kind !== 'spacer' && !options.preserveBlankKind ? 'context' : kind,
-    empty,
-  }
-}
-
 function splitDiffSource(source: unknown) {
   const code = getDisplayCode(source, isLoading.value)
   if (!code)
@@ -177,602 +127,20 @@ function splitDiffSource(source: unknown) {
   return code.split(/\r\n|\n|\r/)
 }
 
-function hasFinalNewline(source: unknown) {
-  return /(?:\r\n|\n|\r)$/.test(String(source ?? ''))
-}
-
-function createMetadataLine(
-  key: string,
-  metadataKind: 'context' | 'removed' | 'added',
-): DiffPreviewLine {
-  return {
-    ...toDiffPreviewLine('No newline at end of file', 'metadata'),
-    key,
-    metadataKind,
-    number: '',
-  }
-}
-
-function appendInlineSourceMetadata(
-  lines: DiffPreviewLine[],
-  originalSource: unknown,
-  modifiedSource: unknown,
-) {
-  const originalMissing = String(originalSource ?? '').length > 0 && !hasFinalNewline(originalSource)
-  const modifiedMissing = String(modifiedSource ?? '').length > 0 && !hasFinalNewline(modifiedSource)
-  if (!originalMissing && !modifiedMissing)
-    return lines
-
-  if (originalMissing)
-    lines.push(createMetadataLine('inline-no-newline-original', 'removed'))
-  if (modifiedMissing)
-    lines.push(createMetadataLine('inline-no-newline-modified', 'added'))
-  return lines
-}
-
-function shouldPreserveSourceBlankDiffKind(lines: string[], index: number) {
-  return !isBlankDiffPreviewLine(lines[index]) || index < lines.length - 1
-}
-
-function isRemovedDiffLine(line: string) {
-  return line.startsWith('-') && !line.startsWith('---')
-}
-
-function isAddedDiffLine(line: string) {
-  return line.startsWith('+') && !line.startsWith('+++')
-}
-
-function hasUnifiedDiffHeaders(lines: string[]) {
-  return lines.some(line => DIFF_HEADER_PREFIXES.some(prefix => line.startsWith(prefix)))
-}
-
-function normalizeLooseDiffBody(body: string, hasHeaders: boolean) {
-  return !hasHeaders && body.startsWith(' ') && !body.startsWith('  ')
-    ? ` ${body}`
-    : body
-}
-
-function isExplicitDiffLanguage() {
-  if (normalizedLanguage.value === 'diff')
-    return true
-
-  const firstLine = String(props.node?.raw ?? '').split(/\r?\n/, 1)[0]?.trim() ?? ''
-  return /^`{3,}\s*diff(?:\s|$)|^~{3,}\s*diff(?:\s|$)/.test(firstLine)
-}
-
-function hasPatchLines(lines: string[]) {
-  const hasRemoved = lines.some(line => isRemovedDiffLine(line))
-  const hasAdded = lines.some(line => isAddedDiffLine(line))
-  return (hasRemoved && hasAdded)
-    || (isExplicitDiffLanguage() && (hasRemoved || hasAdded))
-}
-
-function hasDiffSourcePair() {
-  return props.node?.originalCode != null || props.node?.updatedCode != null
-}
-
-function computeSourceLineMatches(original: string[], modified: string[]): SourceLineMatch[] {
-  const n = original.length
-  const m = modified.length
-  const prefixMatches: SourceLineMatch[] = []
-  let start = 0
-  while (start < n && start < m && original[start] === modified[start]) {
-    prefixMatches.push({ originalIndex: start, modifiedIndex: start })
-    start++
-  }
-
-  const suffixMatches: SourceLineMatch[] = []
-  let originalEnd = n - 1
-  let modifiedEnd = m - 1
-  while (
-    originalEnd >= start
-    && modifiedEnd >= start
-    && original[originalEnd] === modified[modifiedEnd]
-  ) {
-    suffixMatches.unshift({ originalIndex: originalEnd, modifiedIndex: modifiedEnd })
-    originalEnd--
-    modifiedEnd--
-  }
-
-  const middleOriginalLength = originalEnd - start + 1
-  const middleModifiedLength = modifiedEnd - start + 1
-  if (middleOriginalLength <= 0 || middleModifiedLength <= 0)
-    return prefixMatches.concat(suffixMatches)
-  if (isLoading.value)
-    return prefixMatches.concat(suffixMatches)
-
-  const maxCells = 1_500_000
-  if ((middleOriginalLength + 1) * (middleModifiedLength + 1) > maxCells)
-    return prefixMatches.concat(suffixMatches)
-
-  const cols = middleModifiedLength + 1
-  const dp = new Uint32Array((middleOriginalLength + 1) * (middleModifiedLength + 1))
-  for (let i = middleOriginalLength - 1; i >= 0; i--) {
-    for (let j = middleModifiedLength - 1; j >= 0; j--) {
-      const index = i * cols + j
-      if (original[start + i] === modified[start + j]) {
-        dp[index] = dp[(i + 1) * cols + j + 1] + 1
-      }
-      else {
-        const top = dp[(i + 1) * cols + j]
-        const left = dp[i * cols + j + 1]
-        dp[index] = top >= left ? top : left
-      }
-    }
-  }
-
-  const matches: SourceLineMatch[] = []
-  let i = 0
-  let j = 0
-  while (i < middleOriginalLength && j < middleModifiedLength) {
-    if (original[start + i] === modified[start + j]) {
-      matches.push({ originalIndex: start + i, modifiedIndex: start + j })
-      i++
-      j++
-    }
-    else if (dp[(i + 1) * cols + j] >= dp[i * cols + j + 1]) {
-      i++
-    }
-    else {
-      j++
-    }
-  }
-  return prefixMatches.concat(matches, suffixMatches)
-}
-
-function buildInlinePatchPreviewLines(lines: string[]): DiffPreviewLine[] {
-  const result: DiffPreviewLine[] = []
-  let originalLine = 1
-  let modifiedLine = 1
-  const hasHeaders = hasUnifiedDiffHeaders(lines)
-
-  for (const [index, raw] of lines.entries()) {
-    if (raw === NO_NEWLINE_METADATA) {
-      const previousKind = result.at(-1)?.kind
-      const metadataKind = previousKind === 'removed' || previousKind === 'added'
-        ? previousKind
-        : 'context'
-      result.push(createMetadataLine(`inline-no-newline-${index}`, metadataKind))
-    }
-    else if (raw.startsWith('@@')) {
-      const match = raw.match(/^@@\s+-(\d+)(?:,\d+)?\s+\+(\d+)(?:,\d+)?\s+@@/)
-      if (match) {
-        originalLine = Number(match[1])
-        modifiedLine = Number(match[2])
-      }
-      result.push({
-        ...toDiffPreviewLine(raw, 'hunk'),
-        key: `inline-hunk-${index}`,
-        number: '',
-      })
-    }
-    else if (isRemovedDiffLine(raw)) {
-      result.push({
-        ...toDiffPreviewLine(normalizeLooseDiffBody(raw.slice(1), hasHeaders), 'removed', { preserveBlankKind: true }),
-        key: `inline-removed-${index}`,
-        number: originalLine++,
-      })
-    }
-    else if (isAddedDiffLine(raw)) {
-      result.push({
-        ...toDiffPreviewLine(normalizeLooseDiffBody(raw.slice(1), hasHeaders), 'added', { preserveBlankKind: true }),
-        key: `inline-added-${index}`,
-        number: modifiedLine++,
-      })
-    }
-    else {
-      const code = hasHeaders && raw.startsWith(' ') ? raw.slice(1) : raw
-      result.push({
-        ...toDiffPreviewLine(code),
-        key: `inline-context-${index}`,
-        number: modifiedLine,
-      })
-      originalLine++
-      modifiedLine++
-    }
-  }
-
-  return result
-}
-
-function buildInlineSourcePreviewLines(originalSource: unknown, modifiedSource: unknown): DiffPreviewLine[] {
-  const original = splitDiffSource(originalSource)
-  const modified = splitDiffSource(modifiedSource)
-  const matches = computeSourceLineMatches(original, modified)
-  if (matches.length > 0) {
-    const result: DiffPreviewLine[] = []
-    let originalIndex = 0
-    let modifiedIndex = 0
-
-    for (const match of matches) {
-      while (originalIndex < match.originalIndex) {
-        result.push({
-          ...toDiffPreviewLine(original[originalIndex], 'removed', {
-            preserveBlankKind: shouldPreserveSourceBlankDiffKind(original, originalIndex),
-          }),
-          key: `inline-removed-source-${originalIndex}`,
-          number: originalIndex + 1,
-        })
-        originalIndex++
-      }
-      while (modifiedIndex < match.modifiedIndex) {
-        result.push({
-          ...toDiffPreviewLine(modified[modifiedIndex], 'added', {
-            preserveBlankKind: shouldPreserveSourceBlankDiffKind(modified, modifiedIndex),
-          }),
-          key: `inline-added-source-${modifiedIndex}`,
-          number: modifiedIndex + 1,
-        })
-        modifiedIndex++
-      }
-      result.push({
-        ...toDiffPreviewLine(modified[match.modifiedIndex]),
-        key: `inline-context-source-${match.originalIndex}-${match.modifiedIndex}`,
-        number: match.modifiedIndex + 1,
-      })
-      originalIndex = match.originalIndex + 1
-      modifiedIndex = match.modifiedIndex + 1
-    }
-
-    while (originalIndex < original.length) {
-      result.push({
-        ...toDiffPreviewLine(original[originalIndex], 'removed', {
-          preserveBlankKind: shouldPreserveSourceBlankDiffKind(original, originalIndex),
-        }),
-        key: `inline-removed-source-${originalIndex}`,
-        number: originalIndex + 1,
-      })
-      originalIndex++
-    }
-    while (modifiedIndex < modified.length) {
-      result.push({
-        ...toDiffPreviewLine(modified[modifiedIndex], 'added', {
-          preserveBlankKind: shouldPreserveSourceBlankDiffKind(modified, modifiedIndex),
-        }),
-        key: `inline-added-source-${modifiedIndex}`,
-        number: modifiedIndex + 1,
-      })
-      modifiedIndex++
-    }
-
-    return result
-  }
-
-  const result: DiffPreviewLine[] = []
-  let start = 0
-  let originalEnd = original.length - 1
-  let modifiedEnd = modified.length - 1
-
-  while (
-    start <= originalEnd
-    && start <= modifiedEnd
-    && original[start] === modified[start]
-  ) {
-    result.push({
-      ...toDiffPreviewLine(modified[start]),
-      key: `inline-prefix-${start}`,
-      number: start + 1,
-    })
-    start++
-  }
-
-  const suffix: DiffPreviewLine[] = []
-  while (
-    originalEnd >= start
-    && modifiedEnd >= start
-    && original[originalEnd] === modified[modifiedEnd]
-  ) {
-    suffix.unshift({
-      ...toDiffPreviewLine(modified[modifiedEnd]),
-      key: `inline-suffix-${modifiedEnd}`,
-      number: modifiedEnd + 1,
-    })
-    originalEnd--
-    modifiedEnd--
-  }
-
-  for (let index = start; index <= originalEnd; index++) {
-    result.push({
-      ...toDiffPreviewLine(original[index], 'removed', {
-        preserveBlankKind: shouldPreserveSourceBlankDiffKind(original, index),
-      }),
-      key: `inline-removed-source-${index}`,
-      number: index + 1,
-    })
-  }
-
-  for (let index = start; index <= modifiedEnd; index++) {
-    result.push({
-      ...toDiffPreviewLine(modified[index], 'added', {
-        preserveBlankKind: shouldPreserveSourceBlankDiffKind(modified, index),
-      }),
-      key: `inline-added-source-${index}`,
-      number: index + 1,
-    })
-  }
-
-  return result.concat(suffix)
-}
-
-function buildSideBySideSourcePreviewPanes(
-  originalSource: unknown,
-  modifiedSource: unknown,
-): DiffPreviewPane[] {
-  const originalSourceLines = splitDiffSource(originalSource)
-  const modifiedSourceLines = splitDiffSource(modifiedSource)
-  const matches = computeSourceLineMatches(originalSourceLines, modifiedSourceLines)
-  const originalLines: DiffPreviewLine[] = []
-  const modifiedLines: DiffPreviewLine[] = []
-  let originalIndex = 0
-  let modifiedIndex = 0
-  let blockIndex = 0
-
-  const appendChangedBlock = (originalEnd: number, modifiedEnd: number) => {
-    const rowCount = Math.max(originalEnd - originalIndex, modifiedEnd - modifiedIndex)
-    for (let offset = 0; offset < rowCount; offset++) {
-      const nextOriginalIndex = originalIndex + offset
-      const nextModifiedIndex = modifiedIndex + offset
-      originalLines.push(nextOriginalIndex < originalEnd
-        ? {
-            ...toDiffPreviewLine(originalSourceLines[nextOriginalIndex], 'removed', {
-              preserveBlankKind: shouldPreserveSourceBlankDiffKind(originalSourceLines, nextOriginalIndex),
-            }),
-            key: `original-changed-${blockIndex}-${nextOriginalIndex}`,
-            number: nextOriginalIndex + 1,
-          }
-        : {
-            ...toDiffPreviewLine('', 'spacer'),
-            key: `original-spacer-${blockIndex}-${offset}`,
-            number: '',
-          })
-      modifiedLines.push(nextModifiedIndex < modifiedEnd
-        ? {
-            ...toDiffPreviewLine(modifiedSourceLines[nextModifiedIndex], 'added', {
-              preserveBlankKind: shouldPreserveSourceBlankDiffKind(modifiedSourceLines, nextModifiedIndex),
-            }),
-            key: `modified-changed-${blockIndex}-${nextModifiedIndex}`,
-            number: nextModifiedIndex + 1,
-          }
-        : {
-            ...toDiffPreviewLine('', 'spacer'),
-            key: `modified-spacer-${blockIndex}-${offset}`,
-            number: '',
-          })
-    }
-    originalIndex = originalEnd
-    modifiedIndex = modifiedEnd
-    blockIndex++
-  }
-
-  for (const match of matches) {
-    appendChangedBlock(match.originalIndex, match.modifiedIndex)
-    originalLines.push({
-      ...toDiffPreviewLine(originalSourceLines[match.originalIndex]),
-      key: `original-context-${match.originalIndex}-${match.modifiedIndex}`,
-      number: match.originalIndex + 1,
-    })
-    modifiedLines.push({
-      ...toDiffPreviewLine(modifiedSourceLines[match.modifiedIndex]),
-      key: `modified-context-${match.originalIndex}-${match.modifiedIndex}`,
-      number: match.modifiedIndex + 1,
-    })
-    originalIndex = match.originalIndex + 1
-    modifiedIndex = match.modifiedIndex + 1
-  }
-  appendChangedBlock(originalSourceLines.length, modifiedSourceLines.length)
-
-  const originalMissing = String(originalSource ?? '').length > 0 && !hasFinalNewline(originalSource)
-  const modifiedMissing = String(modifiedSource ?? '').length > 0 && !hasFinalNewline(modifiedSource)
-  if (originalMissing || modifiedMissing) {
-    originalLines.push(originalMissing
-      ? createMetadataLine('original-no-newline', 'removed')
-      : {
-          ...toDiffPreviewLine('', 'spacer'),
-          key: 'original-no-newline-spacer',
-          number: '',
-        })
-    modifiedLines.push(modifiedMissing
-      ? createMetadataLine('modified-no-newline', 'added')
-      : {
-          ...toDiffPreviewLine('', 'spacer'),
-          key: 'modified-no-newline-spacer',
-          number: '',
-        })
-  }
-
-  return collapseDiffPanes([
-    {
-      key: 'original',
-      className: 'markstream-pre__diff-pane--original',
-      lines: originalLines,
-    },
-    {
-      key: 'modified',
-      className: 'markstream-pre__diff-pane--modified',
-      lines: modifiedLines,
-    },
-  ])
-}
-
-function resolveDiffCollapseOptions() {
-  const value = props.diffHideUnchangedRegions
-  if (value == null || value === false)
-    return null
-  const options: CodeBlockDiffHideUnchangedRegionsOptions = value === true ? {} : value
-  if (options.enabled === false)
-    return null
-  return {
-    contextLineCount: Math.max(0, Math.floor(options.contextLineCount ?? 2)),
-    minimumLineCount: Math.max(1, Math.floor(options.minimumLineCount ?? 4)),
-  }
-}
-
-function collapseDiffPanes(panes: DiffPreviewPane[]) {
-  const options = resolveDiffCollapseOptions()
-  if (!options || panes.length < 1 || panes.length > 2)
-    return panes
-  if (panes.length === 2 && panes[0].lines.length !== panes[1].lines.length)
-    return panes
-
-  const original = panes[0].lines
-  const modified = panes[1]?.lines
-  let sourceLineCount = original.length
-  while (
-    sourceLineCount > 0
-    && panes.every(pane => pane.lines[sourceLineCount - 1].kind === 'metadata')
-  ) {
-    sourceLineCount--
-  }
-  const isUnchangedRow = (lineIndex: number) => original[lineIndex].kind === 'context'
-    && (
-      modified === undefined
-      || (
-        modified[lineIndex].kind === 'context'
-        && original[lineIndex].code === modified[lineIndex].code
-      )
-    )
-  const collapsedRanges: Array<{ start: number, end: number, first: boolean, last: boolean }> = []
-  let index = 0
-  while (index < sourceLineCount) {
-    const start = index
-    while (index < sourceLineCount && isUnchangedRow(index)) {
-      index++
-    }
-    const end = index
-    const runLength = end - start
-    if (runLength >= options.minimumLineCount) {
-      const hiddenStart = start + (start === 0 ? 0 : options.contextLineCount)
-      const isTerminalRange = end === sourceLineCount
-      const hiddenEnd = end - (isTerminalRange ? 0 : options.contextLineCount)
-      if (hiddenEnd - hiddenStart >= options.minimumLineCount) {
-        collapsedRanges.push({
-          start: hiddenStart,
-          end: isTerminalRange ? original.length : hiddenEnd,
-          first: hiddenStart === 0,
-          last: isTerminalRange,
-        })
-      }
-    }
-    if (index === start)
-      index++
-  }
-
-  if (!collapsedRanges.length)
-    return panes
-
-  return panes.map((pane, paneIndex) => {
-    const lines: DiffPreviewLine[] = []
-    let sourceIndex = 0
-    for (const range of collapsedRanges) {
-      lines.push(...pane.lines.slice(sourceIndex, range.start))
-      lines.push({
-        code: paneIndex === 0 ? 'Unmodified lines' : '',
-        kind: 'collapsed',
-        empty: false,
-        key: `${pane.key}-collapsed-${range.start}-${range.end}`,
-        number: '',
-        collapsedFirst: range.first,
-        collapsedLast: range.last,
-      })
-      sourceIndex = range.end
-    }
-    lines.push(...pane.lines.slice(sourceIndex))
-    return { ...pane, lines }
-  })
-}
-
 const diffPreviewPanes = computed(() => {
   if (!isDiffPreview.value)
     return []
 
-  const hasPatchDiffLines = hasPatchLines(codeLines.value)
-  const hasSourcePair = hasDiffSourcePair()
-  if (isInlineDiffPreview.value) {
-    const lines = hasSourcePair
-      ? appendInlineSourceMetadata(
-          buildInlineSourcePreviewLines(props.node?.originalCode, props.node?.updatedCode),
-          props.node?.originalCode,
-          props.node?.updatedCode,
-        )
-      : buildInlinePatchPreviewLines(codeLines.value)
-
-    return collapseDiffPanes([
-      {
-        key: 'inline',
-        className: 'markstream-pre__diff-pane--inline',
-        lines,
-      },
-    ])
-  }
-
-  if (!hasPatchDiffLines && hasSourcePair) {
-    return buildSideBySideSourcePreviewPanes(
-      props.node?.originalCode,
-      props.node?.updatedCode,
-    )
-  }
-
-  const original = [] as Array<{ code: string, kind: DiffPreviewLineKind, metadataKind?: 'context' | 'removed' | 'added', empty: boolean }>
-  const modified = [] as Array<{ code: string, kind: DiffPreviewLineKind, metadataKind?: 'context' | 'removed' | 'added', empty: boolean }>
-  const hasHeaders = hasUnifiedDiffHeaders(codeLines.value)
-  let previousKind: 'context' | 'removed' | 'added' = 'context'
-
-  for (const raw of codeLines.value) {
-    if (raw === NO_NEWLINE_METADATA) {
-      if (previousKind === 'removed') {
-        original.push(createMetadataLine('original-patch-no-newline', 'removed'))
-        modified.push(toDiffPreviewLine('', 'spacer'))
-      }
-      else if (previousKind === 'added') {
-        original.push(toDiffPreviewLine('', 'spacer'))
-        modified.push(createMetadataLine('modified-patch-no-newline', 'added'))
-      }
-      else {
-        original.push(createMetadataLine('original-patch-no-newline', 'context'))
-        modified.push(createMetadataLine('modified-patch-no-newline', 'context'))
-      }
-    }
-    else if (raw.startsWith('@@')) {
-      original.push(toDiffPreviewLine(raw, 'hunk'))
-      modified.push(toDiffPreviewLine(raw, 'hunk'))
-      previousKind = 'context'
-    }
-    else if (raw.startsWith('-') && !raw.startsWith('---')) {
-      original.push(toDiffPreviewLine(normalizeLooseDiffBody(raw.slice(1), hasHeaders), 'removed', { preserveBlankKind: true }))
-      previousKind = 'removed'
-    }
-    else if (raw.startsWith('+') && !raw.startsWith('+++')) {
-      modified.push(toDiffPreviewLine(normalizeLooseDiffBody(raw.slice(1), hasHeaders), 'added', { preserveBlankKind: true }))
-      previousKind = 'added'
-    }
-    else {
-      const code = hasHeaders && raw.startsWith(' ') ? raw.slice(1) : raw
-      original.push(toDiffPreviewLine(code))
-      modified.push(toDiffPreviewLine(code))
-      previousKind = 'context'
-    }
-  }
-
-  return collapseDiffPanes([
-    {
-      key: 'original',
-      className: 'markstream-pre__diff-pane--original',
-      lines: original.map((line, index) => ({
-        ...line,
-        key: `original-${index}`,
-        number: line.kind === 'metadata' || line.kind === 'spacer' ? '' : index + 1,
-      })),
-    },
-    {
-      key: 'modified',
-      className: 'markstream-pre__diff-pane--modified',
-      lines: modified.map((line, index) => ({
-        ...line,
-        key: `modified-${index}`,
-        number: line.kind === 'metadata' || line.kind === 'spacer' ? '' : index + 1,
-      })),
-    },
-  ])
+  return buildDiffPreviewPanes({
+    code: props.node?.code,
+    hideUnchangedRegions: props.diffHideUnchangedRegions,
+    inline: isInlineDiffPreview.value,
+    language: props.node?.language,
+    loading: isLoading.value,
+    originalCode: props.node?.originalCode,
+    raw: props.node?.raw,
+    updatedCode: props.node?.updatedCode,
+  })
 })
 
 const lineNumberLayoutStyle = computed(() => {
